requirements.txt
fpdf
flask
flask-cors
pandas
matplotlib
PyJWT>=2.0.0
gunicorn
bcrypt
